add: starting course option, refactoring

+++ src/main/java/ru/rsreu/electivecourses/command/DeleteUserCommand.java
@@ -1,8 +1,11 @@
 package ru.rsreu.electivecourses.command;
 
 import ru.rsreu.electivecourses.command.enums.ActionType;
+import ru.rsreu.electivecourses.model.data.Role;
 import ru.rsreu.electivecourses.model.data.User;
+import ru.rsreu.electivecourses.model.data.enums.RoleEnum;
 import ru.rsreu.electivecourses.model.database.dao.AdministratorDAO;
+import ru.rsreu.electivecourses.model.database.dao.RoleDAO;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
@@ -11,16 +14,24 @@ public class DeleteUserCommand extends Command {
     @Override
     public CommandResult execute(HttpServletRequest request) {
         AdministratorDAO adminDAO = (AdministratorDAO) request.getServletContext().getAttribute("administratorDAO");
-        CommandResult commandResult = new CommandResult("/WEB-INF/deleteUser.jsp", ActionType.FORWARD);
-        Long id = Long.valueOf(request.getParameter("id"));
-        boolean deleted = adminDAO.deleteUser(id);
-        if (deleted) {
-            commandResult.addAttribute("result", "Пользователь удалён.");
+        RoleDAO roleDAO = (RoleDAO) request.getServletContext().getAttribute("roleDAO");
+        Long id = Long.valueOf(request.getParameter("delete"));
+        Role role = new Role(id, roleDAO);
+        String roleName = role.getRoleName();
+        CommandResult commandResult;
+        if (!roleName.equals(RoleEnum.ADMIN.getRoleName())) {
+
+            boolean deleted = adminDAO.deleteUser(id);
+            commandResult = new ShowDeletingUserFormCommand().execute(request);
+            if (deleted) {
+                commandResult.addAttribute("result", "Пользователь удалён.");
+            } else {
+                commandResult.addAttribute("result", "Не было удалено пользователей.");
+            }
         } else {
-            commandResult.addAttribute("result", "Не было удалено пользователей.");
+            commandResult = new ShowDeletingUserFormCommand().execute(request);
+            commandResult.addAttribute("result", "Вы не можете удалять администраторов");
         }
-        List<User> users = adminDAO.getAllUsers();
-        commandResult.addAttribute("usersList", users);
         return commandResult;
     }
 }
